requirements.txt
streamlit
fpdf
pypdf
faiss-cpu
langchain
euriai
langchain-community
sentence-transformers
python-dotenv
langchain-text-splitters